refactor(server): move rotation column logic into injectable SecretEncryptionColumnRotationService
Handlers no longer receive secretEncryptionService and logger as call arguments — they call SecretEncryptionColumnRotationService.countNonCurrentRows() / .rotateColumn() and the service holds those collaborators via DI.
