+++ packages/twenty-server/src/database/commands/secret-encryption-rotation/handlers/application-registration-variable-rotation.handler.ts
@@ -1,19 +1,15 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { Repository } from 'typeorm';
 
+import { SecretEncryptionColumnRotationService } from 'src/database/commands/secret-encryption-rotation/services/secret-encryption-column-rotation.service';
 import {
   type SecretEncryptionRotationContext,
   type SecretEncryptionRotationHandler,
   type SecretEncryptionRotationOutcome,
 } from 'src/database/commands/secret-encryption-rotation/types/secret-encryption-rotation-handler.type';
-import {
-  countEncryptedColumnNonCurrentRows,
-  rotateEncryptedColumn,
-} from 'src/database/commands/secret-encryption-rotation/utils/rotate-encrypted-column.util';
 import { ApplicationRegistrationVariableEntity } from 'src/engine/core-modules/application/application-registration-variable/application-registration-variable.entity';
-import { SecretEncryptionService } from 'src/engine/core-modules/secret-encryption/secret-encryption.service';
 
 const ENCRYPTED_COLUMN = 'encryptedValue';
 
@@ -22,14 +18,11 @@ export class ApplicationRegistrationVariableRotationHandler
   implements SecretEncryptionRotationHandler
 {
   readonly siteName = 'application-registration-variable';
-  private readonly logger = new Logger(
-    ApplicationRegistrationVariableRotationHandler.name,
-  );
 
   constructor(
     @InjectRepository(ApplicationRegistrationVariableEntity)
     private readonly applicationRegistrationVariableRepository: Repository<ApplicationRegistrationVariableEntity>,
-    private readonly secretEncryptionService: SecretEncryptionService,
+    private readonly secretEncryptionColumnRotationService: SecretEncryptionColumnRotationService,
   ) {}
 
   async countRemaining({
@@ -37,7 +30,7 @@ export class ApplicationRegistrationVariableRotationHandler
   }: {
     primaryKeyId: string;
   }): Promise<number> {
-    return countEncryptedColumnNonCurrentRows({
+    return this.secretEncryptionColumnRotationService.countNonCurrentRows({
       repository: this.applicationRegistrationVariableRepository,
       primaryKeyId,
       encryptedColumns: [ENCRYPTED_COLUMN],
@@ -47,11 +40,9 @@ export class ApplicationRegistrationVariableRotationHandler
   async rotate(
     context: SecretEncryptionRotationContext,
   ): Promise<SecretEncryptionRotationOutcome> {
-    return rotateEncryptedColumn({
+    return this.secretEncryptionColumnRotationService.rotateColumn({
       ...context,
       repository: this.applicationRegistrationVariableRepository,
-      secretEncryptionService: this.secretEncryptionService,
-      logger: this.logger,
       siteName: this.siteName,
       encryptedColumn: ENCRYPTED_COLUMN,
     });
