+++ packages/twenty-server/src/database/commands/secret-encryption-rotation/handlers/application-variable-rotation.handler.ts
@@ -1,19 +1,15 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { Repository } from 'typeorm';
 
+import { SecretEncryptionColumnRotationService } from 'src/database/commands/secret-encryption-rotation/services/secret-encryption-column-rotation.service';
 import {
   type SecretEncryptionRotationContext,
   type SecretEncryptionRotationHandler,
   type SecretEncryptionRotationOutcome,
 } from 'src/database/commands/secret-encryption-rotation/types/secret-encryption-rotation-handler.type';
-import {
-  countEncryptedColumnNonCurrentRows,
-  rotateEncryptedColumn,
-} from 'src/database/commands/secret-encryption-rotation/utils/rotate-encrypted-column.util';
 import { ApplicationVariableEntity } from 'src/engine/core-modules/application/application-variable/application-variable.entity';
-import { SecretEncryptionService } from 'src/engine/core-modules/secret-encryption/secret-encryption.service';
 
 const ENCRYPTED_COLUMN = 'value';
 const ONLY_SECRET_ROWS_WHERE = `"isSecret" = true`;
@@ -23,12 +19,11 @@ export class ApplicationVariableRotationHandler
   implements SecretEncryptionRotationHandler
 {
   readonly siteName = 'application-variable';
-  private readonly logger = new Logger(ApplicationVariableRotationHandler.name);
 
   constructor(
     @InjectRepository(ApplicationVariableEntity)
     private readonly applicationVariableRepository: Repository<ApplicationVariableEntity>,
-    private readonly secretEncryptionService: SecretEncryptionService,
+    private readonly secretEncryptionColumnRotationService: SecretEncryptionColumnRotationService,
   ) {}
 
   async countRemaining({
@@ -36,7 +31,7 @@ export class ApplicationVariableRotationHandler
   }: {
     primaryKeyId: string;
   }): Promise<number> {
-    return countEncryptedColumnNonCurrentRows({
+    return this.secretEncryptionColumnRotationService.countNonCurrentRows({
       repository: this.applicationVariableRepository,
       primaryKeyId,
       encryptedColumns: [ENCRYPTED_COLUMN],
@@ -47,11 +42,9 @@ export class ApplicationVariableRotationHandler
   async rotate(
     context: SecretEncryptionRotationContext,
   ): Promise<SecretEncryptionRotationOutcome> {
-    return rotateEncryptedColumn({
+    return this.secretEncryptionColumnRotationService.rotateColumn({
       ...context,
       repository: this.applicationVariableRepository,
-      secretEncryptionService: this.secretEncryptionService,
-      logger: this.logger,
       siteName: this.siteName,
       encryptedColumn: ENCRYPTED_COLUMN,
       workspaceIdColumn: 'workspaceId',
