+++ packages/twenty-server/src/database/commands/secret-encryption-rotation/handlers/totp-secrets-rotation.handler.ts
@@ -1,18 +1,14 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { Repository } from 'typeorm';
 
+import { SecretEncryptionColumnRotationService } from 'src/database/commands/secret-encryption-rotation/services/secret-encryption-column-rotation.service';
 import {
   type SecretEncryptionRotationContext,
   type SecretEncryptionRotationHandler,
   type SecretEncryptionRotationOutcome,
 } from 'src/database/commands/secret-encryption-rotation/types/secret-encryption-rotation-handler.type';
-import {
-  countEncryptedColumnNonCurrentRows,
-  rotateEncryptedColumn,
-} from 'src/database/commands/secret-encryption-rotation/utils/rotate-encrypted-column.util';
-import { SecretEncryptionService } from 'src/engine/core-modules/secret-encryption/secret-encryption.service';
 import { TwoFactorAuthenticationMethodEntity } from 'src/engine/core-modules/two-factor-authentication/entities/two-factor-authentication-method.entity';
 
 const ENCRYPTED_COLUMN = 'secret';
@@ -22,12 +18,11 @@ export class TotpSecretsRotationHandler
   implements SecretEncryptionRotationHandler
 {
   readonly siteName = 'totp-secrets';
-  private readonly logger = new Logger(TotpSecretsRotationHandler.name);
 
   constructor(
     @InjectRepository(TwoFactorAuthenticationMethodEntity)
     private readonly twoFactorAuthenticationMethodRepository: Repository<TwoFactorAuthenticationMethodEntity>,
-    private readonly secretEncryptionService: SecretEncryptionService,
+    private readonly secretEncryptionColumnRotationService: SecretEncryptionColumnRotationService,
   ) {}
 
   async countRemaining({
@@ -35,7 +30,7 @@ export class TotpSecretsRotationHandler
   }: {
     primaryKeyId: string;
   }): Promise<number> {
-    return countEncryptedColumnNonCurrentRows({
+    return this.secretEncryptionColumnRotationService.countNonCurrentRows({
       repository: this.twoFactorAuthenticationMethodRepository,
       primaryKeyId,
       encryptedColumns: [ENCRYPTED_COLUMN],
@@ -45,11 +40,9 @@ export class TotpSecretsRotationHandler
   async rotate(
     context: SecretEncryptionRotationContext,
   ): Promise<SecretEncryptionRotationOutcome> {
-    return rotateEncryptedColumn({
+    return this.secretEncryptionColumnRotationService.rotateColumn({
       ...context,
       repository: this.twoFactorAuthenticationMethodRepository,
-      secretEncryptionService: this.secretEncryptionService,
-      logger: this.logger,
       siteName: this.siteName,
       encryptedColumn: ENCRYPTED_COLUMN,
       workspaceIdColumn: 'workspaceId',
